pyproject.toml
[project]
name = "shuzhan-agent"
version = "0.1.0"
description = "数栈智能体工程师 - 离线平台自动化专家"
authors = [{name = "DSY"}]
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shuzhan_agent"]
